Moved validator functions
