+++ src/confluence/confluence.py
@@ -23,9 +23,8 @@ import logging
 from .core.setup import setup_logging
 from .subcommands.merge import parse_args as merge_parse_args
 from .subcommands.merge import main as merge_main
-from .core.validate import parse_args as validate_parse_args
-from .core.validate import main as validate_main
-from .core.validate import validate_dataframe as validate
+from confluence.subcommands.validate import parse_args as validate_parse_args
+from confluence.subcommands.validate import main as validate_main
 
 from confluence import __version__
